Some little changes to deployment
I have added a service to receive the service URL from another live server in order to make this single page application works @@ So dump to use this method but I don't know why Hosting Express App on Cloud Functions does not set cookie on client. And for that reason, I user ngRok to create a tunnel to my localhost express application and everything seems to work fine now!